add GPQA main version.
