+++ nemo_rl/data/eval_datasets/local_math_dataset.py
@@ -1,4 +1,4 @@
-"""Math dataset and its variants."""
+"""Local math dataset."""
 
 from typing import Any, Literal, Optional
 
@@ -8,21 +8,26 @@ from nemo_rl.data import processors
 from nemo_rl.data.interfaces import TaskDataSpec
 
 
-class MathDataset:
+class LocalMathDataset:
     def __init__(
         self,
-        variant: Literal["math_test", "math_500_test"] = "math_test",
+        data_paths: str | list[str],
+        problem_key: str,
+        solution_key: str,
+        name: str,
+        split: Optional[str] = None,
+        file_format: Literal["csv", "json"] = "csv",
         prompt_file: Optional[str] = None,
         system_prompt_file: Optional[str] = None,
     ):
-        ds = load_dataset(
-            "csv",
-            data_files=f"https://openaipublic.blob.core.windows.net/simple-evals/{variant}.csv",
-            split="train",
-        )
+        ds = load_dataset(file_format, data_files=data_paths)
+        if split is not None:
+            ds = ds[split]
+        self._problem_key = problem_key
+        self._solution_key = solution_key
         self.rekeyed_ds = ds.map(self._rekey, remove_columns=ds.column_names)
         self.task_spec = TaskDataSpec(
-            task_name=f"{variant}",
+            task_name=name,
             prompt_file=prompt_file,
             system_prompt_file=system_prompt_file,
         )
@@ -30,6 +35,6 @@ class MathDataset:
 
     def _rekey(self, data: dict[str, Any]):
         return {
-            "problem": data["Question"],
-            "expected_answer": data["Answer"],
+            "problem": data[self._problem_key],
+            "expected_answer": data[self._solution_key],
         }
