+++ nemo_rl/data/eval_datasets/mmlu.py
@@ -4,30 +4,39 @@ from typing import Any, Optional
 
 from datasets import load_dataset
 
+from nemo_rl.data import processors
 from nemo_rl.data.interfaces import TaskDataSpec
 
 
 class MMLUDataset:
-    def __init__(self, prompt_file: Optional[str] = None, system_prompt_file: Optional[str] = None):
-        ds = load_dataset('csv', data_files="https://openaipublic.blob.core.windows.net/simple-evals/mmlu.csv", split='train')
+    def __init__(
+        self,
+        prompt_file: Optional[str] = None,
+        system_prompt_file: Optional[str] = None,
+    ):
+        ds = load_dataset(
+            "csv",
+            data_files="https://openaipublic.blob.core.windows.net/simple-evals/mmlu.csv",
+            split="train",
+        )
         self.rekeyed_ds = ds.map(self._rekey, remove_columns=ds.column_names)
 
         self.task_spec = TaskDataSpec(
-            task_name='MMLU',
+            task_name="MMLU",
             prompt_file=prompt_file,
             system_prompt_file=system_prompt_file,
         )
+        self.processor = processors.multichoice_qa_processor
 
     def _rekey(self, data: dict[str, Any]):
         return {
-            'question': data['Question'],
-            'options': dict(
-                A=data['A'],
-                B=data['B'],
-                C=data['C'],
-                D=data['D'],
+            "question": data["Question"],
+            "options": dict(
+                A=data["A"],
+                B=data["B"],
+                C=data["C"],
+                D=data["D"],
             ),
-            'answer': data['Answer'],
-            'subject': data['Subject'],
+            "answer": data["Answer"],
+            "subject": data["Subject"],
         }
-
