+++ nemo_rl/data/eval_datasets/mmlu_pro.py
@@ -4,28 +4,27 @@ from typing import Any, Optional
 
 from datasets import load_dataset
 
+from nemo_rl.data import processors
 from nemo_rl.data.interfaces import TaskDataSpec
 
 
 class MMLUProDataset:
     def __init__(self, prompt_file: str, system_prompt_file: Optional[str] = None):
-        ds = load_dataset('TIGER-Lab/MMLU-Pro', split='test')
+        ds = load_dataset("TIGER-Lab/MMLU-Pro", split="test")
         self.rekeyed_ds = ds.map(self._rekey, remove_columns=ds.column_names)
 
         self.task_spec = TaskDataSpec(
-            task_name='MMLU-Pro',
+            task_name="MMLU-Pro",
             prompt_file=prompt_file,
             system_prompt_file=system_prompt_file,
         )
+        self.processor = processors.multichoice_qa_processor
 
     def _rekey(self, data: dict[str, Any]):
-        options = {
-            chr(ord('A') + i) : op for i, op in enumerate(data['options'])
-        }
+        options = {chr(ord("A") + i): op for i, op in enumerate(data["options"])}
         return {
-            'question': data['question'],
-            'options': options,
-            'answer': data['answer'],
-            'subject': data['category'],
+            "question": data["question"],
+            "options": options,
+            "answer": data["answer"],
+            "subject": data["category"],
         }
-
